not to discard changes
