+++ tests/test_auto/test_auto.py
@@ -16,24 +16,6 @@ from etna.models import NaiveModel
 from etna.pipeline import Pipeline
 
 
-@pytest.fixture()
-def optuna_storage():
-    yield RDBStorage("sqlite:///test.db")
-    unlink("test.db")
-
-
-@pytest.fixture()
-def trials():
-    class Trial(NamedTuple):
-        user_attrs: dict
-        state: Literal["COMPLETE", "RUNNING", "PENDING"] = "COMPLETE"
-
-    return [
-        Trial(user_attrs={"pipeline": pipeline.to_dict(), "SMAPE_median": i})
-        for i, pipeline in enumerate((Pipeline(NaiveModel(j), horizon=7) for j in range(10)))
-    ]
-
-
 def test_objective(
     example_tsds,
     target_metric=MAE(),
